Clean up: clippy lints, dead code, Default impl
- Apply clippy fixes: is_multiple_of, drop redundant i32 cast, take
  &[RxEphemeris] instead of &Vec in compute_position.
- Remove a dead `if false` debug block in process_samples.
- Drop the never-read UTC decode in subframe 4 page 18 (keep the utc_adj
  flag); remove the now-unused P2_50/getbits2 imports.
- Add Default impl for GnssState.

No behavior change; gpssim fix unchanged (35.6805,139.7650).

+++ src/util.rs
@@ -118,7 +118,7 @@ pub fn getbitu2(buf: &[u8], p1: usize, l1: usize, p2: usize, l2: usize) -> u32 {
 pub fn getbits2(buf: &[u8], p1: usize, l1: usize, p2: usize, l2: usize) -> i32 {
     assert!(l1 + l2 <= 32);
     if getbitu(buf, p1, 1) != 0 {
-        ((getbits(buf, p1, l1) << l2) + getbitu(buf, p2, l2) as i32) as i32
+        (getbits(buf, p1, l1) << l2) + getbitu(buf, p2, l2) as i32
     } else {
         getbitu2(buf, p1, l1, p2, l2) as i32
     }
